+++ app/modules/Application/Mvc/Router/DefaultRouter.php
@@ -8,6 +8,7 @@
 
 namespace Application\Mvc\Router;
 
+use Application\Mvc\Helper\CmsCache;
 use Phalcon\Mvc\Router;
 use Cms\Model\Language;
 
@@ -20,39 +21,38 @@ class DefaultRouter extends Router
     {
         parent::__construct();
 
-        //$this->setDefaultModule('index'); // removed
         $this->setDefaultController('index');
         $this->setDefaultAction('index');
 
-        $this->add('/:module/:controller/:action/:params', array(
+        $this->add('/:module/:controller/:action/:params', [
             'module' => 1,
             'controller' => 2,
             'action' => 3,
             'params' => 4
-        ))->setName('default');
-        $this->add('/:module/:controller', array(
+        ])->setName('default');
+        $this->add('/:module/:controller', [
             'module' => 1,
             'controller' => 2,
             'action' => 'index',
-        ))->setName('default_action');
-        $this->add('/:module', array(
+        ])->setName('default_action');
+        $this->add('/:module', [
             'module' => 1,
             'controller' => 'index',
             'action' => 'index',
-        ))->setName('default_controller');
+        ])->setName('default_controller');
 
     }
 
     public function addML($pattern, $paths = null, $name)
     {
-        $languages = Language::findCachedLanguages();
+        $languages = CmsCache::getInstance()->get('languages');
 
         foreach ($languages as $lang) {
-            $iso = $lang->getIso();
-            if ($lang->getPrimary()) {
+            $iso = $lang['iso'];
+            if ($lang['primary']) {
                 $this->add($pattern, $paths)->setName(self::ML_PREFIX . $name . '_' . $iso);
             } else {
-                $new_pattern = '/' . $lang->getUrl() . $pattern;
+                $new_pattern = '/' . $lang['url'] . $pattern;
                 $paths['lang'] = $iso; // будущее значение константы LANG
                 $this->add($new_pattern, $paths)->setName(self::ML_PREFIX . $name . '_' . $iso);
             }