+++ app/modules/Cms/Model/Translate.php
@@ -1,11 +1,12 @@
 <?php
 /**
-     * @copyright Copyright (c) 2011 - 2014 Aleksandr Torosh (http://wezoom.net)
-     * @author Aleksandr Torosh <webtorua@gmail.com>
-     */
+ * @copyright Copyright (c) 2011 - 2014 Aleksandr Torosh (http://wezoom.net)
+ * @author Aleksandr Torosh <webtorua@gmail.com>
+ */
 
 namespace Cms\Model;
 
+use Application\Mvc\Helper\CmsCache;
 use Phalcon\DI;
 use Phalcon\Mvc\Model;
 
@@ -22,43 +23,15 @@ class Translate extends Model
     public $phrase;
     public $translation;
 
+    public static function translates()
+    {
+        return CmsCache::getInstance()->get('translates');
+    }
+
     public static function findCachedByLangInArray($lang = null)
     {
-        if (!$lang) {
-            $lang = LANG;
-        }
-        $cache = DI::getDefault()->get('cache');
-        $data = $cache->get(self::cacheKey($lang));
-        if (!$data) {
-            $data = self::find(array(
-                'lang = :lang:',
-                'bind' => array(
-                    'lang' => $lang,
-                ),
-            ));
-            if ($data) {
-                $cache->save(self::cacheKey($lang), $data, 300);
-            }
-        }
-
-        $translations = array();
-        if ($data) {
-            foreach ($data as $el) {
-                $translations[$el->getPhrase()] = $el->getTranslation();
-            }
-        }
-        return $translations;
-    }
-
-    public function afterUpdate()
-    {
-        $cache = $this->getDI()->get('cache');
-        $cache->delete(self::cacheKey(LANG));
-    }
-
-    public static function cacheKey($lang)
-    {
-        return HOST_HASH . md5("Translate::findByLang($lang)"); ;
+        $translates = self::translates();
+        return $translates[$lang];
     }
 
     public function findByPhraseAndLang($phrase, $lang = null)
@@ -66,13 +39,13 @@ class Translate extends Model
         if (!$lang) {
             $lang = LANG;
         }
-        $result = self::findFirst(array(
+        $result = self::findFirst([
             'phrase = :phrase: AND lang = :lang:',
-            'bind' => array(
+            'bind' => [
                 'phrase' => $phrase,
-                'lang' => $lang,
-            )
-        ));
+                'lang'   => $lang,
+            ]
+        ]);
         return $result;
     }
 
@@ -139,6 +112,4 @@ class Translate extends Model
     {
         return $this->translation;
     }
-
-
 }